+++ cli/src/main.rs
@@ -1,8 +1,10 @@
 mod chart;
 mod config;
 mod import;
+mod set_start;
 mod start;
 mod summary;
+mod tag;
 mod tags;
 mod time_input;
 
@@ -39,6 +41,12 @@ enum Command {
     #[structopt(name = "tags")]
     Tags(tags::TagsCmd),
 
+    #[structopt(name = "tag")]
+    Tag(tag::Cmd),
+
+    #[structopt(name = "set-start")]
+    SetStart(set_start::Cmd),
+
     #[structopt(name = "import")]
     Import(import::ImportCmd),
 }
@@ -63,6 +71,9 @@ pub enum Error {
     SyncError {
         errors: Vec<RepositoryError<SyncFolderStoreError>>,
     },
+
+    #[snafu(display("Error: {}", source))]
+    GeneralError { source: Box<dyn std::error::Error> },
 }
 
 fn main() {
@@ -124,6 +135,26 @@ fn run() -> Result<(), Error> {
         Command::Summary(subcmd) => subcmd.exec(&timesheet),
         Command::Chart(subcmd) => subcmd.exec(&timesheet),
         Command::Tags(subcmd) => subcmd.exec(&timesheet),
+        Command::Tag(subcmd) => {
+            let patches = subcmd
+                .exec(&timesheet)
+                .map_err(|e| Box::new(e).into())
+                .context(GeneralError {})?;
+            for patch in patches {
+                println!("{}", patch.patch_ref());
+                repo.add_patch(patch).unwrap();
+            }
+        }
+        Command::SetStart(subcmd) => {
+            let patches = subcmd
+                .exec(&timesheet)
+                .map_err(|e| Box::new(e).into())
+                .context(GeneralError {})?;
+            for patch in patches {
+                println!("{}", patch.patch_ref());
+                repo.add_patch(patch).unwrap();
+            }
+        }
     };
 
     // Save which patches this device uses to disk