+++ core/src/repository/timesheet.rs
@@ -58,7 +58,13 @@ impl PatchedTimesheet {
                     continue;
                 }
             };
-            event.add_start(patch_ref.clone(), start_added.time.clone())
+            event.add_start(patch_ref.clone(), start_added.time.clone());
+
+            // Update metadata
+            for parent in start_added.parents() {
+                event.remove_patch_from_latest(&parent);
+            }
+            event.add_patch_to_latest(patch_ref.clone());
         }
         for start_removed in patch.remove_start.iter() {
             let event = match new_state.events.get_mut(&start_removed.event) {
@@ -71,7 +77,14 @@ impl PatchedTimesheet {
                     continue;
                 }
             };
-            event.remove_start(start_removed.patch.clone(), start_removed.time.clone())
+            event.remove_start(start_removed.patch.clone(), start_removed.time.clone());
+
+            // Update metadata
+            event.remove_patch_from_latest(&start_removed.patch);
+            for parent in start_removed.parents() {
+                event.remove_patch_from_latest(&parent);
+            }
+            event.add_patch_to_latest(patch_ref.clone());
         }
 
         for tag_added in patch.add_tag.iter() {
@@ -79,14 +92,27 @@ impl PatchedTimesheet {
                 .events
                 .get_mut(&tag_added.event)
                 .expect("no event for add-tag");
-            event.add_tag(patch_ref.clone(), tag_added.tag.clone())
+            event.add_tag(patch_ref.clone(), tag_added.tag.clone());
+
+            // Update metadata
+            for parent in tag_added.parents() {
+                event.remove_patch_from_latest(&parent);
+            }
+            event.add_patch_to_latest(patch_ref.clone());
         }
         for tag_removed in patch.remove_tag.iter() {
             let event = new_state
                 .events
                 .get_mut(&tag_removed.event)
                 .expect("no event for remove-tag");
-            event.remove_tag(tag_removed.patch.clone(), tag_removed.tag.clone())
+            event.remove_tag(tag_removed.patch.clone(), tag_removed.tag.clone());
+
+            // Update metadata
+            event.remove_patch_from_latest(&tag_removed.patch);
+            for parent in tag_removed.parents() {
+                event.remove_patch_from_latest(&parent);
+            }
+            event.add_patch_to_latest(patch_ref.clone());
         }
 
         for new_event in patch.create_event.iter() {
@@ -95,6 +121,10 @@ impl PatchedTimesheet {
             for tag in new_event.tags.iter().cloned() {
                 event.add_tag(patch_ref.clone(), tag);
             }
+
+            // Update metadata
+            event.add_patch_to_latest(patch_ref.clone());
+
             match new_state.events.insert(new_event.event.clone(), event) {
                 Some(_previous_event) => {
                     errors.push(Error::DuplicateEventId {
@@ -113,14 +143,16 @@ impl PatchedTimesheet {
         }
     }
 
-    pub fn flatten(&self) -> Result<Timesheet, Vec<Error>> {
-        let mut timesheet = Timesheet::new();
+    pub fn flatten(&self) -> Result<Timesheet<'_>, Vec<Error>> {
+        let mut timesheet = Timesheet::new(&self);
         let mut errors = Vec::new();
         let mut event_datetimes_to_refs: BTreeMap<DateTime<Utc>, EventRef> = BTreeMap::new();
         for (event_ref, patched_event) in self.events.iter() {
             match patched_event.flatten() {
                 Ok(event) => {
-                    if let Some(_event_a_tags) = timesheet.insert_event(event.clone()) {
+                    if let Some(_event_a_tags) =
+                        timesheet.event_at_time(event.start().clone(), event_ref.clone())
+                    {
                         errors.push(Error::DuplicateEventTime {
                             event_a: event_datetimes_to_refs[event.start()].clone(),
                             event_b: event_ref.clone(),